requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
eidreader==1.0.8
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.39.0
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyKCS11==1.5.17
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
st-pages==1.0.1
streamlit==1.45.1
streamlit-cookies-controller==0.0.4
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
opencv-python~=4.11.0.86
torch~=2.7.0
ultralytics~=8.3.147
torchvision~=0.22.0
scikit-learn~=1.6.1
tqdm~=4.67.1
shapely~=2.1.1
matplotlib~=3.10.3